refactor(userInputRequester): controller에서 view로 패키지 이동
