+++ src/test/java/racingcar/view/UserInputRequesterTest.java
@@ -1,4 +1,4 @@
-package racingcar.controller;
+package racingcar.view;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -6,7 +6,6 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.io.ByteArrayInputStream;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.view.ErrorMessage;
 
 class UserInputRequesterTest {
